requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
cachelib==0.10.2
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
Flask==2.2.5
Flask-Session==0.4.1
flatbuffers==23.3.3
frozenlist==1.3.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
idna==3.4
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
multidict==6.0.4
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.6
opt-einsum==3.3.0
packaging==23.1
protobuf==4.22.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pymongo==4.3.3
python-dotenv==1.0.0
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow-estimator==2.12.0
tensorflow-macos==2.12.0
termcolor==2.3.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
validate-email==1.3
Werkzeug==2.3.3
wrapt==1.14.1
yarl==1.9.2
